+++ sentry-rails/spec/sentry/rails_spec.rb
@@ -44,7 +44,11 @@ RSpec.describe Sentry::Rails, type: :request do
     it "sets transaction to ControllerName#method" do
       get "/exception"
 
-      expect(event['transaction']).to eq("HelloController#exception")
+      expect(transport.events.last.transaction).to eq("HelloController#exception")
+
+      get "/posts"
+
+      expect(transport.events.last.transaction).to eq("PostsController#index")
     end
 
     it "sets correct request url" do
@@ -153,7 +157,11 @@ RSpec.describe Sentry::Rails, type: :request do
       it "sets transaction to ControllerName#method" do
         get "/exception"
 
-        expect(event['transaction']).to eq("HelloController#exception")
+        expect(transport.events.last.transaction).to eq("HelloController#exception")
+
+        get "/posts"
+
+        expect(transport.events.last.transaction).to eq("PostsController#index")
       end
 
       it "sets correct request url" do