+++ sentry-rails/CHANGELOG.md
@@ -5,6 +5,15 @@
 - Inspect exception cause by default & don't exclude ActiveJob::DeserializationError [#1180](https://github.com/getsentry/sentry-ruby/pull/1180)
   - Fixes [#1071](https://github.com/getsentry/sentry-ruby/issues/1071)
 - Make sentry-rails a Rails engine and provide default job class for async [#1181](https://github.com/getsentry/sentry-ruby/pull/1181)
+- Allow users to configure ActiveJob adapters to ignore [#1256](https://github.com/getsentry/sentry-ruby/pull/1256)
+- Tag `job_id` and `provider_job_id` on ActiveJob events [#1259](https://github.com/getsentry/sentry-ruby/pull/1259)
+- Use another method for post initialization callback [#1261](https://github.com/getsentry/sentry-ruby/pull/1261)
+
+## 4.1.7
+
+- Use env to carry original transaction name [#1255](https://github.com/getsentry/sentry-ruby/pull/1255)
+- Fix duration of tracing event in Rails 5 [#1254](https://github.com/getsentry/sentry-ruby/pull/1254) (by @abcang)
+- Filter out static file transaction [#1247](https://github.com/getsentry/sentry-ruby/pull/1247)
 
 ## 4.1.6
 
@@ -75,4 +84,3 @@ Release test
 ## 0.1.0
 
 First version
-